[WIP] Remove copied and pasted function prototype
The goal of this commit is to make it easier to create a future commit.
That future commit will change the signature of the FindArg() function.
This change makes it so that there’s one less thing that will have to be
updated when FindArg()’s signature is updated.

TODO: Test on:
• Linux
• Windows