[WIP] Remove GetMultiCDPath()
The main motivation behind this change was to make it easier to make a
future change.

There is a side effect to this change. Before this change, mn3_Open()
could be called with a relative path. Now, mn3_Open() must be called
with an absolute path. mn3_Open() was called with an absolute path most
of the time, anyway.

TODO:
• Figure out which platforms to test this on.
• Finish writing commit message.