[WIP] Make mve_Init() not care where movies are stored
Before this change, mve_Init() took two arguments. One of those
arguments was the path to the movies/ directory. mve_Init() would save a
copy of that parameter in the MovieDir array, but there’s no code that
actually does anything with the MovieDir array. In other words, one of
mve_Init()’s arguments was pointless.

This commit removes that pointless argument. The main motivation behind
this commit is to make it easier to make a future commit. That future
commit will make it so that there can be more than one movies/
directory.

TODO:
• Figure out what platforms to test on.
• Proofread and test this commit for the most recent rebase.