Remove unused function
The main motivation behind this change is to make it easier to make a
future commit. That future commit will replace Base_directory with a new
variable. DemoPlayAutoDemo() uses the Base_directory variable. If I
don’t remove DemoPlayAutoDemo(), then I’ll have to update it when
creating that future commit. It’s just easier to remove it.