Revert "Use clang-format with manual trigger only"
This reverts commit 97a22f5bf2c2e16d204acacc4002e5c5b9fb8cdf.

The changes from that commit were only meant to be temporary, anyway.
The commit message itself says “Change it to manual use for the time
being.” The commit message doesn’t indicate when the clang-format Check
workflow should be changed to automatic again, so now seems like as good
a time as any.