Convert table of command-line options into list
Before this change, USAGE.md had an HTML table that contained
documentation for all of Descent 3’s command-line options. A concern was
raised that an HTML table might not be easy to read if someone
downloaded Descent 3 and opened USAGE.md in a plain text editor [1].
This commit alleviates that concern by converting the HTML table into
bulleted lists. Credit for this idea goes to @Lgt2x [2].

The main motivation behind this change is to help get #623 merged.

[1]: <https://github.com/DescentDevelopers/Descent3/pull/623#discussion_r1790331994>
[2]: <https://github.com/DescentDevelopers/Descent3/pull/623#discussion_r1790741430>